IGAPP-1029: Remove widget, when changing city
